[DXP-1660] Fix Maven 3.9.9 build
