+++ core/src/main/java/dev/streamx/cli/command/ingestion/IngestionArgumentConfig.java
@@ -1,6 +1,5 @@
 package dev.streamx.cli.command.ingestion;
 
-import com.fasterxml.jackson.core.JsonParser.Feature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.Option;
@@ -8,31 +7,22 @@ import com.jayway.jsonpath.spi.json.JsonProvider;
 import com.jayway.jsonpath.spi.mapper.JacksonMappingProvider;
 import com.jayway.jsonpath.spi.mapper.MappingProvider;
 import dev.streamx.cli.command.ingestion.publish.payload.PropertyCreatingJacksonJsonNodeJsonProvider;
+import io.quarkus.runtime.Startup;
 import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.enterprise.context.Dependent;
 import java.util.Set;
 
-@Dependent
-public class Config {
+@ApplicationScoped
+@Startup
+public class IngestionArgumentConfig {
 
   private final JsonProvider jsonProvider;
   private final MappingProvider mappingProvider;
 
-  @ApplicationScoped
-  @PayloadProcessing
-  ObjectMapper payloadProcessingObjectMapper() {
-    ObjectMapper objectMapper = new ObjectMapper();
-    objectMapper.enable(Feature.ALLOW_SINGLE_QUOTES);
-
-    return objectMapper;
-  }
-
-  Config(@PayloadProcessing ObjectMapper objectMapper) {
+  public IngestionArgumentConfig(@PayloadProcessing ObjectMapper objectMapper) {
     jsonProvider = new PropertyCreatingJacksonJsonNodeJsonProvider(objectMapper);
     mappingProvider = new JacksonMappingProvider(objectMapper);
     configureDefaults();
   }
-
 
   private void configureDefaults() {
     Configuration.setDefaults(new Configuration.Defaults() {
